# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen laundry - Flutter Test Project",
        "title": "Mewing Laundry API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/customers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Dapatkan daftar pelanggan",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Tambah pelanggan baru",
                "parameters": [
                    {
                        "description": "Customer data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pelanggan berhasil ditambahkan",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Data tidak valid",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Update pelanggan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated customer data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pelanggan berhasil diperbarui",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Pelanggan tidak ditemukan",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Delete pelanggan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pelanggan berhasil dihapus",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Pelanggan tidak ditemukan",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login dengan email dan password, return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login pengguna",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Email atau password salah",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    }
}